fix duplicate extended css and js in EntryContentBase (HtmlHelpers.cs) and clean up code
