--- src/Foundation.Cms/CalloutContentAlignments.cs
@@ -1,10 +1,0 @@
-namespace Foundation.Cms
-{
-    public static class CalloutContentAlignments
-    {
-        public const string None = "None";
-        public const string Left = "Left";
-        public const string Right = "Right";
-        public const string Center = "Center";
-    }
-}
