+++ src/Foundation/Features/Category/StandardCategory.cs
@@ -8,8 +8,7 @@ using System.ComponentModel.DataAnnotations;
 
 namespace Foundation.Features.Category
 {
-    [ContentType(
-        GUID = "A9BBD7FC-27C5-4718-890A-E28ACBE5EE26",
+    [ContentType(GUID = "A9BBD7FC-27C5-4718-890A-E28ACBE5EE26",
         DisplayName = "Standard Category",
         Description = "Used to categorize content")]
     public class StandardCategory : CategoryData, IFoundationContent
