+++ src/Foundation/Features/Category/StandardCategoryViewModel.cs
@@ -4,7 +4,12 @@ namespace Foundation.Features.Category
 {
     public class StandardCategoryViewModel : ContentViewModel<StandardCategory>
     {
-        public StandardCategoryViewModel() { }
-        public StandardCategoryViewModel(StandardCategory category) : base(category) { }
+        public StandardCategoryViewModel() 
+        { 
+        }
+
+        public StandardCategoryViewModel(StandardCategory category) : base(category) 
+        { 
+        }
     }
 }
