+++ src/Foundation/Features/Header/HeaderViewModelFactory.cs
