+++ src/Foundation/Features/Locations/Blocks/FilterContinentsBlock.cs
@@ -27,7 +27,10 @@ namespace Foundation.Features.Locations.Blocks
         [Display(Name = "All condition text")]
         public virtual string AllConditionText { get; set; }
 
-        public ITypeSearch<LocationItemPage.LocationItemPage> AddFilter(ITypeSearch<LocationItemPage.LocationItemPage> query) => query.TermsFacetFor(x => x.Continent);
+        public ITypeSearch<LocationItemPage.LocationItemPage> AddFilter(ITypeSearch<LocationItemPage.LocationItemPage> query)
+        {
+            return query.TermsFacetFor(x => x.Continent);
+        }
 
         public ITypeSearch<LocationItemPage.LocationItemPage> ApplyFilter(ITypeSearch<LocationItemPage.LocationItemPage> query, NameValueCollection filters)
         {
