+++ src/Foundation/Features/Locations/LocationListPage/LocationListViewModel.cs
@@ -7,7 +7,9 @@ namespace Foundation.Features.Locations.LocationListPage
 {
     public class LocationListViewModel : ContentViewModel<LocationListPage>
     {
-        public LocationListViewModel(LocationListPage currentPage) : base(currentPage) { }
+        public LocationListViewModel(LocationListPage currentPage) : base(currentPage) 
+        {
+        }
 
         public GeoCoordinate MapCenter { get; set; }
         public IGeolocationResult UserLocation { get; set; }
