+++ src/Foundation/Features/Locations/TagPage/TagsCarouselViewModel.cs
