+++ src/Foundation/Features/MyAccount/CreditCard/CreditCardPage.cs
@@ -1,7 +1,7 @@
 using EPiServer.DataAbstraction;
 using EPiServer.DataAnnotations;
 using Foundation.Features.Shared;
-using Foundation.Features.Shared.Descriptors;
+using Foundation.Features.Shared.EditorDescriptors;
 using Foundation.Infrastructure;
 using System.ComponentModel.DataAnnotations;
