+++ src/Foundation/Features/MyOrganization/Budgeting/BudgetingPage.cs
@@ -1,6 +1,6 @@
 using EPiServer.DataAnnotations;
 using Foundation.Features.Shared;
-using Foundation.Features.Shared.Descriptors;
+using Foundation.Features.Shared.EditorDescriptors;
 using Foundation.Infrastructure;
 
 namespace Foundation.Features.MyOrganization.Budgeting
