+++ src/Foundation/Features/MyOrganization/Organization/OrganizationPage.cs
@@ -1,7 +1,7 @@
 using EPiServer.DataAbstraction;
 using EPiServer.DataAnnotations;
 using Foundation.Features.Shared;
-using Foundation.Features.Shared.Descriptors;
+using Foundation.Features.Shared.EditorDescriptors;
 using Foundation.Infrastructure;
 
 namespace Foundation.Features.MyOrganization.Organization
