+++ src/Foundation/Features/Preview/PreviewModel.cs
@@ -6,8 +6,7 @@ namespace Foundation.Features.Preview
 {
     public class PreviewModel : ContentViewModel<FoundationPageData>
     {
-        public PreviewModel(FoundationPageData currentPage, IContent previewContent)
-            : base(currentPage)
+        public PreviewModel(FoundationPageData currentPage, IContent previewContent) : base(currentPage)
         {
             PreviewContent = previewContent;
             Areas = new List<PreviewArea>();
