+++ src/Foundation/Features/Shared/EditorDescriptor/ColorPickerEditorDescriptor.cs
@@ -3,7 +3,7 @@ using EPiServer.Shell.ObjectEditing.EditorDescriptors;
 using System;
 using System.Collections.Generic;
 
-namespace Foundation.Features.Shared.Descriptors
+namespace Foundation.Features.Shared.EditorDescriptors
 {
     [EditorDescriptorRegistration(TargetType = typeof(string), UIHint = "ColorPicker")]
     public class ColorPickerEditorDescriptor : EditorDescriptor
