+++ src/Foundation/Infrastructure/PowerSlices/CommerceSlices.cs
