+++ src/Foundation/Infrastructure/PowerSlices/ContentSlices.cs
