+++ src/Foundation/Infrastructure/PowerSlices/GeneralSlices.cs
